requirements.txt
# Web Frameworks
fastapi
uvicorn[standard]
python-multipart

# Frontend
streamlit
requests

# Utilities & AI
python-dotenv
google-genai

# Data Handling & PDF
pydantic
pypdf